Refactor -> TransactionRequest.class renamed in Transaction.class and moved to Entity package User.class moved to entity package Added Tests for UserService
