+++ src/main/java/com/softserve/academy/antifraudsystem6802/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.softserve.academy.antifraudsystem6802.controller;
 
-import com.softserve.academy.antifraudsystem6802.model.User;
+import com.softserve.academy.antifraudsystem6802.model.entity.User;
 import com.softserve.academy.antifraudsystem6802.model.request.RequestLock;
 import com.softserve.academy.antifraudsystem6802.model.request.RoleRequest;
 import com.softserve.academy.antifraudsystem6802.model.response.ResponseDelete;
