+++ src/test/java/com/softserve/academy/antifraudsystem6802/controllerTest/TransactionControllerTest.java
@@ -1,11 +1,10 @@
 package com.softserve.academy.antifraudsystem6802.controllerTest;
 
 import com.softserve.academy.antifraudsystem6802.model.Role;
-import com.softserve.academy.antifraudsystem6802.model.User;
+import com.softserve.academy.antifraudsystem6802.model.entity.User;
 import com.softserve.academy.antifraudsystem6802.model.request.RequestLock;
 import com.softserve.academy.antifraudsystem6802.model.request.RoleRequest;
 import com.softserve.academy.antifraudsystem6802.service.UserService;
-import org.aspectj.lang.annotation.Before;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -18,7 +17,6 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -333,7 +331,7 @@ public class TransactionControllerTest {
     }
 
     @Test
-    public void transactionPostBlacklList() throws Exception {
+    public void transactionPostBlackList() throws Exception {
         String trP1 = "{\n" +
                 "  \"amount\": 1000,\n" +
                 "  \"ip\": \"192.168.1.67\",\n" +
